Site/requirements.txt
Flask 
Flask-SQLAlchemy 
Flask-Mail 
flask 
bcrypt 
flask_sqlalchemy
flask_mail 
python-dotenv
